+++ api/src/game/game.service.spec.ts
@@ -3,6 +3,7 @@ import { Test } from '@nestjs/testing';
 import { AnalyticsService } from '../analytics/analytics.service';
 import { EventRewardsService } from '../event-rewards/event-rewards.service';
 import { MembersService } from '../members/members.service';
+import { PlayerDtoService } from '../player/player-dto.service';
 import { PlayerService } from '../player/player.service';
 import { PlayerPropertyService } from '../player-property/player-property.service';
 
@@ -38,6 +39,10 @@ describe('GameService', () => {
         },
         {
           provide: AnalyticsService,
+          useValue: {},
+        },
+        {
+          provide: PlayerDtoService,
           useValue: {},
         },
       ],
