+++ api/src/game/game.service.ts
@@ -4,12 +4,11 @@ import { logger } from 'firebase-functions';
 import { EventRewardsService } from '../event-rewards/event-rewards.service';
 import { Member } from '../members/entities/members.entity';
 import { MembersService } from '../members/members.service';
-import { PlayerSettingService } from '../player/player-setting.service';
+import { PlayerDtoService } from '../player/player-dto.service';
 import { PlayerService } from '../player/player.service';
 import { PlayerPropertyService } from '../player-property/player-property.service';
 
 import { GameResetPlayerProperty } from './dto/game-reset-player-property';
-import { PlayerDto } from './dto/player.dto';
 import { PointInfoDto } from './dto/point-info.dto';
 @Injectable()
 export class GameService {
@@ -19,7 +18,7 @@ export class GameService {
     private readonly membersService: MembersService,
     private readonly eventRewardsService: EventRewardsService,
     private readonly playerPropertyService: PlayerPropertyService,
-    private readonly playerSettingService: PlayerSettingService,
+    private readonly playerDtoService: PlayerDtoService,
   ) {}
 
   getOK(): string {
@@ -123,7 +122,8 @@ export class GameService {
   async resetPlayerProperty(gameResetPlayerProperty: GameResetPlayerProperty): Promise<void> {
     const { steamId, useMemberPoint } = gameResetPlayerProperty;
 
-    const player = (await this.findPlayerDtoBySteamIds([steamId.toString()]))[0];
+    // FIXME 不使用playerDto，直接使用playerService获取player
+    const player = await this.playerDtoService.findPlayerDtoBySteamId(steamId);
 
     if (!player) {
       throw new BadRequestException();
@@ -139,6 +139,7 @@ export class GameService {
         memberPointTotal: -resetPlayerPropertyMemberPoint,
       });
     } else {
+      // FIXME 不使用playerDto，直接使用playerService中的方法计算。计算考虑做成helper
       const resetPlayerPropertySeasonPoint = player.seasonNextLevelPoint;
       if (player.seasonPointTotal < resetPlayerPropertySeasonPoint) {
         throw new BadRequestException();
@@ -150,44 +151,5 @@ export class GameService {
 
     // 重置玩家属性
     await this.playerPropertyService.deleteBySteamId(steamId);
-  }
-
-  async findPlayerDtoBySteamId(steamId: number): Promise<PlayerDto> {
-    const players = await this.findPlayerDtoBySteamIds([steamId.toString()]);
-    return players[0];
-  }
-
-  // TODO 移动到playerModule
-  async findPlayerDtoBySteamIds(ids: string[]): Promise<PlayerDto[]> {
-    const players = (await this.playerService.findByIds(ids)) as PlayerDto[];
-    for (const player of players) {
-      const properties = await this.playerPropertyService.findBySteamId(+player.id);
-      if (properties) {
-        player.properties = properties;
-      } else {
-        player.properties = [];
-      }
-      const setting = await this.playerSettingService.getPlayerSettingOrGenerateDefault(player.id);
-      player.playerSetting = setting;
-
-      const seasonPoint = player.seasonPointTotal;
-      const seasonLevel = this.playerService.getSeasonLevelBuyPoint(seasonPoint);
-      player.seasonLevel = seasonLevel;
-      player.seasonCurrrentLevelPoint =
-        seasonPoint - this.playerService.getSeasonTotalPoint(seasonLevel);
-      player.seasonNextLevelPoint = this.playerService.getSeasonNextLevelPoint(seasonLevel);
-
-      const memberPoint = player.memberPointTotal;
-      const memberLevel = this.playerService.getMemberLevelBuyPoint(memberPoint);
-      player.memberLevel = memberLevel;
-      player.memberCurrentLevelPoint =
-        memberPoint - this.playerService.getMemberTotalPoint(memberLevel);
-      player.memberNextLevelPoint = this.playerService.getMemberNextLevelPoint(memberLevel);
-      player.totalLevel = seasonLevel + memberLevel;
-
-      const usedLevel = player.properties.reduce((prev, curr) => prev + curr.level, 0);
-      player.useableLevel = player.totalLevel - usedLevel;
-    }
-    return players;
   }
 }
